test/generator/requirements.txt
psycopg2-binary==2.9.7
faker==18.9.0
python-dotenv==1.0.1